+++ src/faah/installer/__init__.py
@@ -1,0 +1,1 @@
+"""Install faah shell hooks and managed config files."""